CiteSource: A package to compare sources of citation records

The CiteSource package supports evidence aggregation by helping with the processing of results of various searches in different sources. It allows to deduplicate results while retaining meta-data on where those results were found and then enables users to compare the contribution of different sources

See also

Author

Maintainer: Trevor Riley trevor.riley@noaa.gov (ORCID)

Authors:

Kaitlyn Hair kaitlyn.hair@ed.ac.uk (ORCID)

Lukas Wallrich lukas.wallrich@gmail.com (ORCID)

Matthew Grainger matthewjamesgrainger@gmail.com (ORCID)

Sarah Young sarahy@andrew.cmu.edu (ORCID)

Chris Pritchard chris.pritchard@ntu.ac.uk (ORCID)

Neal Haddaway nealhaddaway@gmail.com (ORCID)

Other contributors:

Martin Westgate (Author of included synthesisr fragments) [copyright holder]

Eliza Grames (Author of included synthesisr fragments) [copyright holder]